commandHauls: extract scraping to modules
