+++ src/Scrape/command.js
@@ -1,0 +1,38 @@
+import { Command } from '/twcheese/src/Models/Command.js';
+import { parseArrival } from '/twcheese/src/Scrape/time.js';
+import { scrapeResources } from '/twcheese/src/Scrape/res.js';
+
+/**
+ *	scrapes a command page for the command info and returns it as a Command object
+ *	@param	{HTMLDocument} gameDoc	the page generated by game.php?screen=command_info&id=x&type=own
+ *	@return {Command}
+ */
+function scrapeCommand(gameDoc) {
+    var command = new Command();
+
+    try {//note: being lazy - catching exception thrown for returning scouts and outgoing troops instead of checking for them
+        var content = $(gameDoc).find('#content_value').get()[0];
+
+        var arrivalCell = content.getElementsByTagName('table')[0].rows[6].cells[1];
+        command.arrival = parseArrival($(arrivalCell).text());
+
+        var resCell = content.getElementsByTagName('table')[2].rows[0].cells[1];
+        var haul = scrapeResources(resCell);
+        command.timber = haul.timber;
+        command.clay = haul.clay;
+        command.iron = haul.iron;
+
+        var haulText = resCell.innerHTML;
+        if (haulText.search('\\|') !== -1) {
+            haulText = haulText.substring(haulText.search('\\|') + 7);
+            let performance = haulText.split('/');
+            command.haulCapacity = parseInt(performance[1]);
+        }
+    }
+    catch (e) {
+    }
+
+    return command;
+};
+
+export { scrapeCommand }
