+++ src/ToolSetup/commandHauls.js
@@ -27,6 +27,8 @@ import { userConfig } from '/twcheese/src/Util/UserConfig.js';
 import { TwCheeseDate } from '/twcheese/src/Models/TwCheeseDate.js';
 import { Command } from '/twcheese/src/Models/Command.js';
 import { ProgressMonitor } from '/twcheese/src/Models/ProgressMonitor.js';
+import { scrapeCommand } from '/twcheese/src/Scrape/command.js';
+import { scrapePageNumber } from '/twcheese/src/Scrape/pagination.js';
 
 if (!twcheese)
     var twcheese = {};
@@ -42,53 +44,6 @@ twcheese.images = {
     popupBorder: 'graphic/popup/border.png',
     servant: 'graphic/paladin_new.png',
     loadingSpinner: 'graphic/throbber.gif'
-};
-
-/*==== scraper functions ====*/
-
-/**
- *	scrapes a command page for the command info and returns it as a twcheese_Command object
- *	@param	gameDoc:HTMLDocument	the page generated by game.php?screen=command_info&id=x&type=own
- *	@return command:twcheese_Command	an object representing the command.
- */
-twcheese.scrapeCommand = function (gameDoc) {
-    var command = new Command();
-
-    try {//note: being lazy - catching exception thrown for returning scouts and outgoing troops instead of checking for them
-        var content = $(gameDoc).find('#content_value').get()[0];
-
-        var arrivalCell = content.getElementsByTagName('table')[0].rows[6].cells[1];
-        command.arrival = twcheese.parseArrival($(arrivalCell).text());
-
-        var resCell = content.getElementsByTagName('table')[2].rows[0].cells[1];
-        var haul = twcheese.scrapeResources(resCell);
-        command.timber = haul.timber;
-        command.clay = haul.clay;
-        command.iron = haul.iron;
-
-        var haulText = resCell.innerHTML;
-        if (haulText.search('\\|') !== -1) {
-            haulText = haulText.substring(haulText.search('\\|') + 7);
-            let performance = haulText.split('/');
-            command.haulCapacity = parseInt(performance[1]);
-        }
-    }
-    catch (e) {
-    }
-
-    return command;
-};
-
-/**
- * @param {string} text formatted the way tw does it. e.g. "Jun 12, 2019  15:36:23:000"
- * @return {TwCheeseDate}
- */
-twcheese.parseArrival = function (text) {
-    // note: some worlds have milliseconds disabled
-    let expr = /(\D{3}) (\d{1,2}), (\d{4})  (\d{2}):(\d{2}):(\d{2}):?(\d{3})?/;
-    let [, monthName, day, year, hours, minutes, seconds, millis] = text.match(expr);    
-    let month = TwCheeseDate.monthNumber(monthName);
-    return TwCheeseDate.newServerDate(year, month, day, hours, minutes, seconds, millis || 0);
 };
 
 /**
@@ -117,35 +72,6 @@ twcheese.requestDocumentBody = async function (targetUrl) {
         }
         xmlhttp.send("");
     });
-};
-
- /**
-  * @param {HTMLElement} resourcesContainer an element containing timber/clay/iron amounts
-  * @return {Object} {timber: timberAmount, clay: clayAmount, iron: ironAmount}
-  */
-twcheese.scrapeResources = function (resourcesContainer) {
-    // remove grey periods used as thousands separators
-    let $res = $(resourcesContainer).clone().remove('.grey'); 
-
-    let resAmount = function(resIconCssClass) {
-        // note: sometimes, if the res amount is 0, the game excludes it (and its icon) instead of showing 0
-        let icon = $res.find('span.' + resIconCssClass).get(0);
-        return icon ? parseInt($(icon.nextSibling).text()) : 0;
-    }
-
-    return {
-        timber: resAmount('wood'),
-        clay: resAmount('stone'),
-        iron: resAmount('iron')
-    };
-};
-
-twcheese.scrapePageNumber = function() {
-    let currentPage = $('#paged_view_content').children('table:eq(0)').find('strong').html();
-    if (currentPage && !currentPage.includes('all')) {
-        return parseInt(currentPage.match(/\d+/)[0]);
-    }
-    return null;
 };
 
 /*==== widgets ====*/
@@ -311,7 +237,7 @@ twcheese.createPillagingStatsWidget = function(commands, collapsed) {
         startOfHour = startOfHour.addHours(1);
     }
 
-    let pageNumber = twcheese.scrapePageNumber();
+    let pageNumber = scrapePageNumber();
     let pageInfo = pageNumber ? `from Page ${pageNumber}` : '';
 
     let toggleIconSrc = collapsed ? twcheese.images.plus : twcheese.images.minus;
@@ -440,7 +366,7 @@ twcheese.appendHaulColsToCommandsTable = async function (progressMonitor) {
         }
 
         let commandUrl = firstCell.getElementsByTagName('a')[0].href;
-        let command = twcheese.scrapeCommand(await twcheese.requestDocumentBody(commandUrl));
+        let command = scrapeCommand(await twcheese.requestDocumentBody(commandUrl));
         let commandType = $(firstCell).find('.own_command').data('command-type');
         if (commandType === 'return') {            
             returningCommands.push(command);
